argus/prototype/requirements.txt
websocket-client==1.8.0
keyboard==0.13.5